fixed decoder issue, changed some settings
basically added many stuff, like the collate_fn now returns 3 items, inputs, labels and images
inputs are captions[:-1] and labels are captions[1:]
we feed the inputs to the model and use labels for cross entropy
but it seems this is not mandetory, as we should be able to train the model using
captions it seems. but overfitting and bias would be an issue.

after this change we no more get 99+ accuracy in neither hidden_state or input mode
the hidden_state mode gave me the best output with 67.43% val accuracy. the text wasnt garbage
and it actually creates meaningful captions.
I havent tried the input yet. I also need to add that I initialized c0 like h0 with the image features
previously we would only initialize the h0 with image features for some reason. it seems to me it might
have been a good addition. I need to test that more and make sure if that is the case.

we arent doing much for training really, i could use more nonlinearity in encoder, use relu, bn, or use
layernorm after lstm, use dropouts etc, but we havent. I just want to get an understanding of how this
works.