good settings for training
this seems a good starting point for training.
seems the image encoder part is very important
when we see the model creates gramatically correct yet semantically wrong description
it means the language modeling part is developed properly but the image features part
or the utilization of it is not good enough. so try to work on the image encoder part
to see improvements in the description quality.