lower timesteps_t from 400 to 200
when it was 400, I used to get pure noise even at 1920 epoch! the loss
was going down though, so it may very well have been the sampling issue
not being able to properly create image that long!? or shor? (cuz we only
tried 10 images for 400 timesteps? (i dont think so, cuz at least we had to
get the first image more clearer, but it seemed the model couldnt propelry
denoise!?