add gen_images method which doesnt work! read notes
notes:
added a new method called gen_images to our model but somehow
doesnt work and causes an error! its either a bug for model or
im missing something here.
aside from that I tested the mnist dataset with the current changes
(disabled tanh, disabled bn for final_conv) with 200 and 400 timesteps
and got prefect results early on.
theres no more noise spots on the image, they are smooth and prefect.
I noticed with a low timestep_num like 100, the generation takes longer
to generate good images, when we used t=400, as early as 20/40 epochs we had
prefect numbers formed, however, when t=100, even til epoc 80 we have giberish
yet clean/smooth images (this shows the importance of timesteps. the higher
the timesteps, the btter the results the earlier!