fix gen_image method and somemore changes
the gen_image method failed with the weird error:
"RuntimeError: GET was unable to find an engine to execute this computation"
only becasue of out of bound index, that is the timestep that I was using should
have been less. that is we cant use timestep of 800 if our num_timesteps=800 because
the the betas tensor are 0 based, so its only 0-799. when we loop in the _sample()
method the range(0, num_timesteps) is actually between 0 and 799 not 800!
so when I was trying to use 800 it would fail but pytorch wouldnt catch it and instead
create the wrong error message.
on top of that, the implementation of gen_image was wrong, we have to have a loop to generate
an image. the best looking image is the one at 0ths step (we count in reverse so we start from 799
and loop til we reach 0, this is when we get our best image)
I change this and it now works.
we need to refactor this and display_sample later on, but for now I want to leave it be

I also added a nother architecture to test, but disabled it. I also tried some other changes but
nothing worked.
I also tried using a smaller start_eps for betas, also tried not rescaling -1,1 and using only .ToTensor(),
but nothing changed for cifar. I guess we might be overfitting! lets see