Add new architecture which drastically improves the result.
without any conditioning or timestep usage, a simple base hourglass architcture
did amazing!
the images are now vibrant and natural looking, so it means when the images are dark/grimish
the model is underfitting. and by underfitting it doesnt mean its lacking parameters,
it just means, the discriminative power is not there, the parameters are not beng utilized
efficiently and model can not learn properly or as fast! so a proper architecture massively
affects this!