use base architecture without timestep_embedding
it converges much faster, is more vibrant compared to when using timestep embedding
but still not as fast or as good/vibrant as the new architecture

this implies the new architecture is much better, (its smaller 23m vs 54m )
and also the timeemedding at least like how im doing is not helping at all and makes everything grimish/gray!
next we can use the new sampling, and then use the improvement in new architecture samples (furiour,etc) and see
how it affects the final result.

so to racap:
1.new architecture gives us much better results much quicker
2.new archityecture doesnt use timestep embedding or any conditioning
3.timestep emebediing makes the results grimish/gray/darker
4.a small model as small as 23m works better than our 54m model.
5.