readded dropou to new arch and add more explanation
basically dropout doesnt play a role here, the output were just as good as before
and i didnt notice any kind of difference neither in image quality nor in the loss

I guess the scheduler is probably playing a significant role here (aside from the timeembedding which was crucial for getting vibrant good images)
there are two remaining factors we need to test
1.loss
2.noise scheduler

so next we will be working on the old architecture, and start with changing the loss and scheduler
and document their effects. and then we will focus on the old architecture and
remove the old unet architecture and use an hour glass one like the one in new architetcure and then
chnage the way timeembedding is introduced and also add class conditioning as well later on.

so to recap, we test with
1.loss
2.noise scheduler
3.change architecture to hourglass form
4.use a different method for adding time embedding to the architecture.