Add a loss compatible with our linear scheduler
the previous loss didnt fare well even up to 1600 epochs
it was very noisy, and we really couldnt make a damn thing in the images (they are
availble in /imgs_gen_20240412_21_25_50 directory)
so we changed it to suit our linear sampler and here we are training .
by the way, even with our new test which is being run up to the current epoch which is 440
we are still seeing a lot of noise in the image, images are just pretty noisy and loss really
isnt improving per epoch as before, its around 0.2269 /0.2265/0.2251 and jumps up and down!
we'll see how it goes!