add explanation for loss update(works very good:) )
the images are vibrant and its much much better, the images have deformed objects
though, but the overall picture looks real and vibrant. need more epochs to see how
it turns out ultimately. also at 660 epochs we are at 0.1158 loss, around 1000 0.1114
and the quality became much better (less deformed). (seems the lr decrease at 1000 decreased it
but after 1200 loss stopped improving and increased from 0.1100 to .1104 then 0.1106 to 0.1108 in 1240!
signaling its not improving much until 1260 which became 0.1102 (so its fluctuating at best I guess!)

next we need to test with weights and see how that affects the result and whether it
helps in our case or not. (how we formulate the weight may very well matter so keep that in mind)