test with betas_end=0.02 with low timesteps like 200 so it works like 0.008 and timesteps=500 combo
tested with new scheduler/optimizer and while the convergence was fsater and images were more visible this time
we seem to need to add more dropout ratio, we still overfit it seems. as the lr decays more the rate of change decreases
so we need to get to a low loss at early epochs to have good formed images, as the lr decays, the images rarely change & if
by that time we dont have well formed visible images, we wont get any improvement. at lower lrs, the quality and sharpness of images
improves, but not the overall form of images.

anyway, I decided to try and see if theres a connection between timesteps and betas values, and there is. if we use high betas like
betas_end=0.02, we need to use a lower timesteps, otherwise it results in white/black blobs in the image. the loss decreases but the
images doesnt improve, this signals us that the sampling part is doing sth wrong, either its too much iteration or too few iterations
or maybe the values are out of range for somereasons, resulting in weird image generations. if loss gives us nans, etc thats a loss issue
and we need to check the betas value range and other operations involved. in our case it was betas range which was very large! coupled with
high timesteps, it created all of those issues (instable loss, bad image generations (white/blackblos for images ) etc

now in this test, we set betas_end=0.02 and lowered the timesteps to 200 to have good image generation and so far its working well
need to test higher values of timesteps and see where it stops improving