fix fp16 issue (unscaled and scaled param mismatch issue)
the unscaling part had to be done at the very end,when we are done
doing backward passes, just before we do the optimizer step(),
otherwise it will mess up the gradient accumulation
(some will be scaled and some will be unscaled)
therefore when we take a step, it will result in a huge update, and either
we diverge asap or get nans. (we diverged! cuz I added plr later, after
norm calculation and this made a whole mess. only after moving the norm calculation